Factor out a Packages mixin for Participant
