+++ gratipay/models/participant/email.py
@@ -9,7 +9,7 @@ from psycopg2 import IntegrityError
 
 import gratipay
 from gratipay.exceptions import EmailAlreadyVerified, EmailTaken, CannotRemovePrimaryEmail
-from gratipay.exceptions import EmailNotVerified, TooManyEmailAddresses, EmailNotOnFile, NoPackages
+from gratipay.exceptions import EmailNotVerified, TooManyEmailAddresses, EmailNotOnFile
 from gratipay.security.crypto import constant_time_compare
 from gratipay.utils import encode_for_querystring
 
@@ -190,31 +190,6 @@ class Email(object):
         return nonce
 
 
-    def start_package_claims(self, c, nonce, *packages):
-        """Takes a cursor, nonce and list of packages, inserts into ``claims``
-        and returns ``None`` (or raise :py:exc:`NoPackages`).
-        """
-        if not packages:
-            raise NoPackages()
-
-        # We want to make a single db call to insert all claims, so we need to
-        # do a little SQL construction. Do it in such a way that we still avoid
-        # Python string interpolation (~= SQLi vector).
-
-        extra_sql, values = [], []
-        for p in packages:
-            extra_sql.append('(%s, %s)')
-            values += [nonce, p.id]
-        c.run('INSERT INTO claims (nonce, package_id) VALUES' + ', '.join(extra_sql), values)
-        self.app.add_event( c
-                          , 'participant'
-                          , dict( id=self.id
-                                , action='start-claim'
-                                , values=dict(package_ids=[p.id for p in packages])
-                                 )
-                               )
-
-
     def set_primary_email(self, email, cursor=None):
         """Set the primary email address for the participant.
         """
@@ -279,20 +254,6 @@ class Email(object):
             return VERIFICATION_SUCCEEDED, packages, paypal_updated
 
 
-    def get_packages_claiming(self, cursor, nonce):
-        """Given a nonce, return :py:class:`~gratipay.models.package.Package`
-        objects associated with it.
-        """
-        return cursor.all("""
-            SELECT p.*::packages
-              FROM packages p
-              JOIN claims c
-                ON p.id = c.package_id
-             WHERE c.nonce=%s
-          ORDER BY p.name ASC
-        """, (nonce,))
-
-
     def save_email_address(self, cursor, address):
         """Given an email address, modify the database.
 
@@ -315,31 +276,6 @@ class Email(object):
         """, (self.id, address))
         if not self.email_address:
             self.set_primary_email(address, cursor)
-
-
-    def finish_package_claims(self, cursor, nonce, *packages):
-        """Create teams if needed and associate them with the packages.
-        """
-        if not packages:
-            raise NoPackages()
-
-        package_ids, teams, team_ids = [], [], []
-        for package in packages:
-            package_ids.append(package.id)
-            team = package.get_or_create_linked_team(cursor, self)
-            teams.append(team)
-            team_ids.append(team.id)
-        review_url = self.app.project_review_repo.create_issue(*teams)
-
-        cursor.run('DELETE FROM claims WHERE nonce=%s', (nonce,))
-        cursor.run('UPDATE teams SET review_url=%s WHERE id=ANY(%s)', (review_url, team_ids,))
-        self.app.add_event( cursor
-                          , 'participant'
-                          , dict( id=self.id
-                                , action='finish-claim'
-                                , values=dict(package_ids=package_ids)
-                                 )
-                               )
 
 
     def get_email(self, address, cursor=None, and_lock=False):
